Get rid of ucharVectorPtr, use ucharVector as dictionary key
This change gives two benefits: using ucharVector as key gives us
half of the comparisons when looking for an entry in the dictionary,
and our life will be much easier when saving/loading a binary file.
Plus, some binary->string conversions on FindPassword were removed.

To avoid using custom hash functions, and to still have a reference
to our hash for saving purposes, dictionary is now reverted to
regular std::map.